+++ accounts/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import BaseUser
+
+# Register your models here.
+admin.site.register(BaseUser)
