+++ hairsalon/admin.py
@@ -1,9 +1,8 @@
 from django.contrib import admin
-from .models import Salon, BaseUser, SalonStaff, News
+from .models import Salon, SalonStaff, News
 
 # Register your models here.
 
-admin.site.register(BaseUser)
 admin.site.register(News)
 admin.site.register(SalonStaff)
 admin.site.register(Salon)
